organized directories, fixed css bug
