+++ src/App.jsx
@@ -3,7 +3,7 @@ import { Login } from "./components/auth/Login"
 import { Register } from "./components/auth/Register"
 import { Authorized } from "./views/Authorized"
 import { ApplicationViews } from "./views/ApplicationViews"
-import { UserProvider } from "./contexts/UserIdContext"
+import { UserProvider } from "./customReact/contexts/UserIdContext"
 
 export const App = () => {
