+++ src/components/Doc/Doc.css
@@ -53,7 +53,7 @@
 }
 
 .title {
-    font-size: clamp(1rem, 2rem, 2rem);
+    font-size: clamp(15px, 25px, 25px);
     font-weight: bold;
     text-align: center;
 }
