+++ src/components/myDocs/MyDocs.jsx
@@ -1,13 +1,13 @@
 import { useContext, useEffect, useState } from "react"
 import { deleteDocById, getDocsByUserId } from "../../services/docsService"
-import { UserContext } from "../../contexts/UserIdContext"
+import { UserContext } from "../../customReact/contexts/UserIdContext"
 import { Doc } from "../Doc/Doc"
 import "./MyDocs.css"
 
 export const MyDocs = () => {
     const [myDocs, setMyDocs] = useState([])
     const [deleteId, setDeleteId] = useState(0)
-    const {userId} = useContext(UserContext)
+    const { userId } = useContext(UserContext)
 
     useEffect(() => {
         if (userId > 0) {
@@ -15,16 +15,22 @@ export const MyDocs = () => {
                 setMyDocs(res)
             })
         }
+    }, [userId])
 
+    useEffect(() => {
         if (deleteId > 0) {
-            deleteDocById(deleteId)
+            deleteDocById(deleteId).then(() => {
+                getDocsByUserId(userId).then((res) => {
+                    setMyDocs(res)
+                })
+            })
         }
     }, [userId, deleteId])
 
     return (
         <div className="myDocs-container">
             {myDocs.map(docInfo => {
-                return <Doc key={docInfo.id} docInfo={docInfo} setDeleteId={setDeleteId} deleteId={deleteId}/>
+                return <Doc key={docInfo.id} docInfo={docInfo} setDeleteId={setDeleteId} deleteId={deleteId} />
             })}
         </div>
     )
